refactor: normalize package entry points and version pinning
Packages disagreed on where their root declarations live (`dist/index.d.ts`
vs `dist/src/index.d.ts`), two packages had no `exports` map at all, and two
used caret ranges despite the repo-wide `saveExact: true` policy.

Types layout — every package now resolves its root types to `dist/index.d.ts`,
sitting next to the `dist/index.js` it describes:

- storage: `tsconfig.build.json` rootDir `.` -> `src` (it had no sibling
  `types/` dir, so this is config-only).
- logger, cache: moved the sibling `types/` sources under `src/types/`, matching
  the layout storage already used, so declarations root at `src`. For cache the
  `@` alias is repointed from `.` to `src`, which keeps every existing
  `@/types/*` import working unchanged.
- workflows: `vite-plugin-dts` now uses `entryRoot: "src"`.

`dist/**` no longer carries `*.test.d.ts` for cache, workflows and i18n-utils,
which brings them in line with ai/logger/storage/jobs. i18n-utils gains a
`tsconfig.build.json` (the ai/logger/storage pattern) so build-time declaration
emit is separate from `typecheck`, which still covers tests.

Exports maps — added to the two packages that lacked one. Both are consumed via
deep source subpaths, so the maps preserve exactly that: `email` exposes `.`
plus `./src/*`, and `i18n-utils` exposes its built root plus `./src/*`. The five
packages using bare-condition exports now nest them under a `"."` key
(behaviour-identical) so all maps share one shape, and entry paths are
`./`-prefixed throughout.

Version pinning — replaced caret ranges with exact pins in i18n-utils (glob,
object-hash, @types/node, @types/object-hash, tsx) and workflows (rimraf,
typescript, vite-plugin-dts). Every one already resolved to the version pinned
elsewhere in the repo, so the lockfile diff is specifier-only with no version
changes. survey-ui's `peerDependencies` ranges are intentionally left as ranges.

Also fixes two dead entry points the audit surfaced, both of which pointed at
files that have never existed:

- js-core declared `import: "./dist/index.js"`, but its Vite config only emits
  UMD (`dist/index.umd.cjs`). Now points at the real artifact; the redundant
  `typesVersions` wildcard is dropped and `main`/`types` added.
- @formbricks/types declared `main`/`exports["."]` as `./index.ts`, a file with
  no history in this repo and zero importers. Removed; all real usage goes
  through the `./*` pattern.

i18n-utils' bundle filenames change from `index.es.js`/`index.cjs.js` to
`index.js`/`index.cjs` to match every other package; nothing referenced the old
names.

Verified: `pnpm build`, `pnpm typecheck`, `pnpm test` and `pnpm lint` all pass
from a forced (uncached) run, `publint` is clean for workflows, and a script
asserts every declared `main`/`module`/`types`/`source`/`exports` target
resolves to an existing file. Lint warnings are unchanged from base.
