+++ packages/ai/package.json
@@ -22,8 +22,10 @@
     "dist"
   ],
   "exports": {
-    "types": "./dist/index.d.ts",
-    "import": "./dist/index.js"
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    }
   },
   "scripts": {
     "clean": "rimraf .turbo node_modules coverage dist",
