+++ packages/cache/src/service.test.ts
@@ -1,9 +1,9 @@
 import { beforeEach, describe, expect, test, vi } from "vitest";
 import { logger } from "@formbricks/logger";
-import type { RedisClient } from "../types/client";
-import { ErrorCode } from "../types/error";
-import type { CacheKey } from "../types/keys";
 import { CacheService } from "./service";
+import type { RedisClient } from "./types/client";
+import { ErrorCode } from "./types/error";
+import type { CacheKey } from "./types/keys";
 
 // Mock logger
 vi.mock("@formbricks/logger", () => ({
