+++ packages/email/package.json
@@ -4,8 +4,15 @@
   "private": true,
   "description": "Email templates for Formbricks with React Email preview server",
   "type": "module",
-  "main": "src/index.ts",
-  "types": "src/index.ts",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./src/*": [
+      "./src/*.ts",
+      "./src/*.tsx"
+    ]
+  },
   "scripts": {
     "dev": "email dev --port 3456",
     "build": "pnpm typecheck",
