+++ packages/i18n-utils/package.json
@@ -6,9 +6,17 @@
   "homepage": "https://formbricks.com",
   "license": "MIT",
   "description": "Utility functions for multilingual support in Formbricks projects",
-  "main": "dist/index.cjs.js",
-  "module": "dist/index.es.js",
-  "types": "dist/index.d.ts",
+  "main": "./dist/index.cjs",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js",
+      "require": "./dist/index.cjs"
+    },
+    "./src/*": "./src/*.ts"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/formbricks/formbricks"
@@ -20,7 +28,8 @@
     "localization"
   ],
   "files": [
-    "dist"
+    "dist",
+    "src"
   ],
   "scripts": {
     "go": "vite build --watch --mode dev",
@@ -29,22 +38,22 @@
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "typecheck": "tsc --noEmit",
-    "build": "tsc && vite build",
+    "build": "tsc --project tsconfig.build.json && vite build",
     "test": "vitest run",
     "test:coverage": "vitest run --coverage",
     "test:watch": "vitest",
     "scan-translations": "tsx src/scan-translations.ts"
   },
   "dependencies": {
-    "glob": "^13.0.6",
-    "object-hash": "^3.0.0"
+    "glob": "13.0.6",
+    "object-hash": "3.0.0"
   },
   "devDependencies": {
     "@formbricks/config-typescript": "workspace:*",
     "@formbricks/eslint-config": "workspace:*",
-    "@types/node": "^25.4.0",
-    "@types/object-hash": "^3.0.6",
-    "tsx": "^4.21.0",
+    "@types/node": "25.4.0",
+    "@types/object-hash": "3.0.6",
+    "tsx": "4.21.0",
     "typescript": "5.9.3",
     "vite": "7.3.5",
     "vitest": "4.1.6"
