+++ packages/js-core/package.json
@@ -19,16 +19,12 @@
   "files": [
     "dist"
   ],
+  "main": "./dist/index.umd.cjs",
+  "types": "./dist/index.d.ts",
   "exports": {
-    "types": "./dist/index.d.ts",
-    "import": "./dist/index.js",
-    "require": "./dist/index.umd.cjs"
-  },
-  "typesVersions": {
-    "*": {
-      "*": [
-        "./dist/index.d.ts"
-      ]
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.umd.cjs"
     }
   },
   "scripts": {
