+++ packages/storage/tsconfig.build.json
@@ -6,9 +6,9 @@
     "isolatedModules": false,
     "noEmit": false,
     "outDir": "dist",
-    "rootDir": "."
+    "rootDir": "src"
   },
   "exclude": ["src/**/*.test.ts", "src/**/*.spec.ts"],
   "extends": "./tsconfig.json",
-  "include": ["src", "types"]
+  "include": ["src"]
 }
