+++ packages/surveys/package.json
@@ -11,10 +11,10 @@
     "url": "https://github.com/formbricks/formbricks"
   },
   "sideEffects": false,
-  "source": "src/index.ts",
-  "main": "dist/index.umd.cjs",
-  "module": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "source": "./src/index.ts",
+  "main": "./dist/index.umd.cjs",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
   "files": [
     "dist"
   ],
