+++ packages/workflows/package.json
@@ -6,16 +6,16 @@
   "type": "module",
   "exports": {
     ".": {
-      "types": "./dist/src/index.d.ts",
+      "types": "./dist/index.d.ts",
       "import": "./dist/index.js"
     },
     "./server": {
-      "types": "./dist/src/server/index.d.ts",
+      "types": "./dist/server/index.d.ts",
       "import": "./dist/server.js"
     }
   },
   "main": "./dist/index.js",
-  "types": "./dist/src/index.d.ts",
+  "types": "./dist/index.d.ts",
   "scripts": {
     "build": "vite build && publint",
     "clean": "rimraf .turbo node_modules coverage dist",
@@ -34,10 +34,10 @@
     "@vitest/coverage-v8": "4.1.6",
     "js-yaml": "4.3.0",
     "publint": "0.3.21",
-    "rimraf": "^6.1.2",
-    "typescript": "^5.9.3",
+    "rimraf": "6.1.3",
+    "typescript": "5.9.3",
     "vite": "7.3.5",
-    "vite-plugin-dts": "^4.5.4",
+    "vite-plugin-dts": "4.5.4",
     "vitest": "4.1.6"
   }
 }
